vm_tools: support tools disk from arbitrary DLC
This allows passing an arbitrary DLC, from which its tools_disk.img will
be provided to a VM. This allows for custom third-party VMs to use the
same tools disk Termina does and also makes it possible for other
Crostini-like VMs to use a shared tools_disk.img instead of duplicating
them.

This also removes Concierge's fallback mechanism for starting Termina if
no DLC is provided since this change makes it otherwise impossible to
run. According to crbug/953544 this should be safe as DLC is a hard
requirement for Crostini now.

BUG=b:189120779
TEST=deployed to hatch and ensured bruschetta started with the specified
tools disk

Change-Id: I0547a97d8769bdd143ceb65ef94f99bb0a9072ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2918786
Tested-by: Mike Gerow <gerow@google.com>
Reviewed-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Mike Gerow <gerow@google.com>