shill: Reset mobile operator info when destroying capability
Currently, the home and serving provider info are being reset when
the SIM is replaced. This has a race condition with OnOperatorChanged
and some information is reset after the data was already received in
OnOperatorChanged. Reset the provider info when destroying the
capability to ensure we have fresh data when creating the capability the
next time a SIM card is inserted.

BUG=b:189095310
TEST=test_that dood suite:cellular_ota --fast

Change-Id: Ia72a19b76b600866a49355baec0fb65b0c00334a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2919496
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>