cryptohome: Move instantiation of AuthBlock to VaultKeyset
Move EncryptVaultKeyset() and DecryptVaultKeyset() to VaultKeyset
together with other dependent functions.

This CL is part of the effort to extend AuthBlocks to include smartcards
and other possible future asynchronous authentication methods. We need
to reduce the AuthBlock call stack before changing the API. Here we
reduce the Crypto call step and move the instantiation of the AuthBlock
to VaultKeyset.

BUG=chromium:1213647
TEST=Unit tests pass.
TEST=tast run DUT hwsec* cryptohome*

Change-Id: If17909c3d1df8998b58a537d5b4f7b5bdd578846
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2930819
Tested-by: Betul Soysal <betuls@google.com>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: Betul Soysal <betuls@google.com>