diagnostics: Create test env from mojo object in system fetcher test
Add functions to create the test env in the system fetcher test. This is
for reducing the lines of the test and for migrating to system info v2.
The test env is created based on |expected_system_info| variable. So the
tests won't need to set or delete the files every time. They only need to
set the expected value and re-init the test env.

BUG=b:188508561
TEST=FEATURES=test emerge-$BOARD diagnostics

Change-Id: I5b9c624eb5d9edd9030ff8f8147df765fb264f70
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2922443
Tested-by: Chung-Sheng Wu <chungsheng@chromium.org>
Reviewed-by: Meng-Huan Yu <menghuan@chromium.org>
Reviewed-by: Yilin Yang (kerker) <kerker@chromium.org>
Commit-Queue: Chung-Sheng Wu <chungsheng@chromium.org>