cryptohome: Fix a couple of bugs that AuthFactors discovered
Bug#1:
When AuthFactor was introduced, it surfaced the fact that KeyData wasnt
being copied to VKK (which we also discovered in another OOBE bug).
Since we know that we always write KeyData to credentials, lets always
write the KeyData to VK.

Bug#2:
AuthSession does not use the fact that an AuthFactor is authenticated or
not to set kAuthStatusAuthenticated.

BUG=chromium:1204663
TEST=tast test - cryptohome.AuthSessionMount.regular_mount

Change-Id: I1382d7630784d961028247141382151f16ac9af2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2937438
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>
Auto-Submit: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>