arc: Fix arcvm-boot-notification-server crash
This change modifies arcvm-boot-notification-server so that, when it is
sending props to the guest, it will continue accepting connections until
it is able to write the props successfully.

This is to address a race between the previous and current VM that
happens when Chrome restarts while the first VM is in its early stages
of boot up. Chrome will not kill the previous VM when shutting down, so
it will continue booting until arcvm-boot-notification-client connect().
However, by the time the previous VM reaches connect(), a new
arcvm-boot-notification-server has started. arcvm-boot-notification-server
only expects one connect request from the guest, but it will fail
because the previous VM is killed before the server can accept(). In
order to allow the current VM to connect and boot successfully, we need to
accept() again.

BUG=b:188450841
TEST=To reproduce crash, apply gpaste/5685150860967936 patch to Chromium.
 Apply this change to arcvm-boot-notification-server, and verify that it
 does not crash

Change-Id: I4d8a8cd27b2d23fc08b737b79ff2fd80e52f49b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2939009
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Cherie Cheung <cherieccy@google.com>
Tested-by: Willie Koomson <wvk@google.com>
Commit-Queue: Yusuke Sato <yusukes@chromium.org>