ml: add space to text_suggester test case
Previously the prediction test cases for the TextSuggester
sharedlib did not end with a space char. The shared library
currently returns prediction suggestions without a preceding
space char when the last char in the text given to the library
does not end with a space char.

This was causing an issue with consumers of the service as
suggestions in this case would be concatenated with the current
text.

This CL adds a space char to the text_suggester test cases
in preparation for a new uprev of the sharedlib to fix the
issue.

BUG=chromium:1146266
TEST=cros_run_unit_tests --board=${BOARD} --packages chromeos-base/ml

Change-Id: Ia7444de8a804cf2f53b397ae48f0c8ec09b340fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2943690
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Tested-by: Curtis McMullan <curtismcmullan@chromium.org>
Commit-Queue: Darren Shen <shend@chromium.org>