arc: libvda: Add support for configuring the bitrate mode to libvda.
This CL adds support for changing the bitrate mode and associated peak
bitrate to the libvda encoder. All bitrate-related variables are packed
into a new Bitrate structure similar to the changes made to the Chrome
VEA in crrev.com/c/2986204.

The supported bitrate modes are:
- VBR (Variable bitrate)
- CBR (Constant bitrate)

Note: the bitrate mode is currently not piped through to the Chrome
encoder yet, as updating the mojo bindings is a WIP.

BUG=b:190336806,b:181514834
TEST=tast run DUT arc.VideoEncodeAccel.h264_192p_i420_vm

Cq-Depend: chromium:2944317
Change-Id: I03e8f2f31b528292e995aa4cfcef9ebda3c4d00d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2946469
Tested-by: David Staessens <dstaessens@chromium.org>
Commit-Queue: David Staessens <dstaessens@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>