init: Add logging for systemd-tmpfiles.
This logs errors from systemd-tmpfiles to /run/tmpfiles.log which is
moved to /var/log after chromeos_startup finishes and is logged to the
syslog after system-services starts. The logs are also appended to the
clobber.log when the stateful partition is clobbered.

Also adds the /proc path to pre-startup since it was not being covered.

BUG=b:189503059
TEST=manual check that the log is created, copied to the right location,
  and present in /var/log/messages

Change-Id: Ib636d49676ee8d2f4e619b55389c3733c83853cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2945073
Tested-by: Allen Webb <allenwebb@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Allen Webb <allenwebb@google.com>