cryptohome: Move Hkdf() call to RecoveryCrypto
Hashing the publisher_dh/destination_dh is an implementation detail of
the recovery crypto.
E.g. if in the future we will change it to compute HKDF on the X
coordinate instead of on the whole point, CryptohomeRecoveryAuthBlock
doesn't need to know about this.

Make `hkdf_salt` values in RecoveryCrypto empty.
Make `hkdf_info` values for recovery key and mediator share constant.

BUG=b:180711640
TEST=USE=test emerge-${BOARD} cryptohome

Change-Id: I6d4b2ffd9de639d825da373e4215c57b5677754b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2944879
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Tested-by: Anastasiia N <anastasiian@chromium.org>
Commit-Queue: Anastasiia N <anastasiian@chromium.org>