ml: extract delegate classes from model and graph_executor
The current model_impl and graph_executor_impl are both impl of mojo
interfaces, and runs in a async way. This is unnecessary when ml-service
runs as a dbus service rather than a mojo service.

This CL extracts two delegate classes from model_impl and
graph_executor_impl which run tensorflow-related stuff synchronously,
i.e building an interpreter, executing a graph. They are used by the
model_impl and graph_executor_impl and the changes are transparent
to the users of ml-service. They can also make it easier when ml-service
runs as a dbus service.

BUG=chromium:1217852
TEST=pass the unittests

Change-Id: Ibdb3f6af1537b5fbfd15187d54a180d17db4eabf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2948024
Tested-by: Xinglong Luan <alanlxl@chromium.org>
Commit-Queue: Xinglong Luan <alanlxl@chromium.org>
Reviewed-by: Andrew Moylan <amoylan@chromium.org>