rmad: Use runtime_probe to get probed components
Integrate with runtime_probe to probe the components on the device. When
entering `ComponentsRepair` state, we first probe the device and report
which components are detected. We only need Yes/No answers here, instead
of the detailed list of components.

For devices that don't have runtime_probe enabled, we'll need to
manually deploy the probe statements to the device for this to work. We
can also ask runtime_probe team to enable it on certain platforms, such
as b:190584056

BUG=b:187196023
TEST=emerge-volteer chromeos-base/system_api rmad;
TEST=manual test on DUT

Cq-Depend: chromium:2951310
Change-Id: Ie518924fb26dd6f3f22d1517ab7db63d2f281541
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2951351
Tested-by: Cheng-Han Yang <chenghan@chromium.org>
Reviewed-by: Yilin Yang (kerker) <kerker@chromium.org>
Reviewed-by: Gene Chang <genechang@google.com>
Reviewed-by: Yun-Kai Lin <kevinptt@chromium.org>
Commit-Queue: Cheng-Han Yang <chenghan@chromium.org>