vm_tools: Turn free_bytes into a DIY optional field
This is needed as the default value (0) is also a valid value, we need
to distinguish the cases and (as evidenced by the related bug) making
sure we set it in every path is hard. Use a separate boolean field
instead of anything fancier for backcompat, we seem to be on an older
version of proto3 without optional fields anyway.

BUG=chromium:1216305
TEST=CQ

Change-Id: Ice6b77ba636dbb51c15bb44312dfca7be2495b5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2952730
Commit-Queue: Nicholas Verne <nverne@chromium.org>
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Reviewed-by: Fergus Dall <sidereal@google.com>
Tested-by: Nicholas Verne <nverne@chromium.org>