dlp: Add tests for DlpAdaptor.
The added tests check correctness of processing
file access request for different scenarious when
the file is known by DLP database or not and there is
DLP policy that might restrict it or not.

Also some possible bugs are fixed (due to missing DB).

BUG=chromium:1169882
TEST=cros_workon_make --board=${BOARD} chromeos-base/dlp --test

Change-Id: I0525bf1ae715a2f798b3fbecf2bfa3e8b6b6939c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2953440
Commit-Queue: Sergey Poromov <poromov@chromium.org>
Tested-by: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Aya Elsayed <ayaelattar@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>