shill: Translate QMI APN error to shill error
When ModemManager fails to connect with an APN when using a QMI modem,
the error reported by ModemManager is not known to shill. Adding a
specific translation from the ModemManager error `CallFailed` to the
shill error `kInvalidApn` will allow shill to retry different APNs when
a connection fails.

BUG=b:190725498
TEST=Enter an invalid APN in the UI with a verizon SIM and verify that
shill fails to connect with that APN and falls back to the good APN
`vzwinternet`.

Change-Id: Ida3dc863bd4354d337bad9deed2e5794ad3215bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2952894
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Auto-Submit: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>