hermes: Set state as inactive after deleting pending profile
Chrome expects all pending profiles to have the state kPending.
There is a transitory period when a profile is in installed_profiles_ as
well as pending_profiles_. During this period, maintain the state as
pending. Mark the profile as inactive once the profile is no longer in
pending_profiles_.

BUG=b:190193212
TEST=unit tests

Change-Id: I49c0510b5f0cc70b7bb6bf5b73a0b9d2721e1cac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2961871
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Pavan Holla <pholla@google.com>
Tested-by: Pavan Holla <pholla@google.com>