ml: Process-isolate SODA model.
Multi-process support for ml-service was added prior to this CL
(see go/cros-ml-multiprocess).

This CL causes the SODA model to be run in a dedicated worker process.

Tested on an amd64 device (eve) but not tested on any arm/arm64
devices because SODA is not yet supported on these.

BUG=chromium:1220271
TEST=on device (eve).
TEST=Live Caption enabled and SODA triggered via YouTube video.
TEST=worker processes observed via `ps`.

Cq-Depend: 2967689
Change-Id: I02f1c5d5de46c8e305260038a19f5d143146133c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2967846
Reviewed-by: Honglin Yu <honglinyu@chromium.org>
Commit-Queue: Amanda Deacon <amandadeacon@chromium.org>
Tested-by: Amanda Deacon <amandadeacon@chromium.org>