patchpanel: Skip apply_local_source_mark if source bits are set
This patch adds a skip rule at the beginning of apply_local_source_mark
chain if a packet already has those bits set. This can avoid that chain
overwrite the vpn policy and source bits on the packet, and will be
useful if the "owner" of the packet and the intended source mark does
not match, (e.g., packets from wireguard kernel socket).

BUG=b:190683881
TEST=tast run ${DUT} network.VPNConnect.wireguard (this test itself is
  still WIP)
TEST=unit_tests
TEST=checked that apply_local_source_mark contains
  `-A apply_local_source_mark -m mark ! --mark 0x0/0x3f00 -j RETURN`
  and saw traffic counters for vpn increasing when pinging wg gateway.

Change-Id: I3463df89cb798c32784b71e04487e7e1ad2c2c4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2966439
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>