cryptohome: extend pcr of TPM and CSME together.
As far as cryptohome is concerned, the PCR values of interest should be
consistent.

As for testing, ATM there is no good way to observe that the PCR-binding
does perform integrity check when performing operations by CSME.
However, for the existing cr50 case, it is verified that the added code
doesn't have side effect.

BUG=b:190621192
TEST=unittests.
TEST=tast run $IP hwsec.LockToSingleUserMountUntilReboot

Change-Id: Ibebe917a4674c1e75b8fabb0def56eb43035ba91
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3005734
Tested-by: Leo Lai <cylai@google.com>
Commit-Queue: Leo Lai <cylai@google.com>
Reviewed-by: Andrey Pronin <apronin@chromium.org>