arc: libvda: Add new GAVD backend implementing mojo::VideoDecoderClient
This CL adds a new GAVD backend that implements the new
mojo::VideoDecoderClient based on the Chrome media::VideoDecoder
interface. This allows LibVDA to directly interface with the
media::VideoDecoder-based Chrome video decoders, as the
media::VideoDecodeAccelerator interface has been deprecated.

Note: The event pipe between crosvm and libvda is currently using an
interface which is loosely based on the deprecrated
media::VideoDecodeAccelerator interface. We should update the pipe in
the future, but since we still have to maintain the GAVDA backend for
now this would require adaptations to the GAVDA backend. Once all
devices have been switched to the new GAVD backend we can simply remove
the GAVDA backend before updating the event pipe.

BUG=b:189278506
TEST=arc.VideoDecodeAccel.h264_vm

Change-Id: Ie02122eaf7234fb700697ba9a9b7cc93825eaf31
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3020341
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Tested-by: David Staessens <dstaessens@chromium.org>
Commit-Queue: David Staessens <dstaessens@chromium.org>