os_install_service: improve install error logging
Changed an error log from "{:?}" to "{}" and updated the relevant use
of thiserror::Error so that details of the nested error are
logged. The practical effect of this is that instead of getting the
error log on a single line with a bunch of "\n"s, you get each line of
the nested error on its own line.

BUG=b:190858576
TEST=Run the installer in a way it will be sure to fail, e.g. by
     putting an "exit 1" in chromeos-install. Tail the log in
     /var/log/os_install_service and verify the error output is good.

Change-Id: I59795ea78eacb1ded40d4306f6e52fd91593575b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2965465
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>