crosh: Enable wireguard commands
Enable the wireguard commands in crosh, except for the following cases:
a) wireguard is not supported on the device. This is done by checking if
   the value of SupportedVPNTypes property in Manager contains
   "wireguard".
b) VPN is disabled on the device. This is done by checking if the value
   of ProhibitedTechnologies property in Manager contains vpn.
c) The device is a managed device. This is done by checking the output
  of `cryptohome --action=install_attributes_get --name=enterprise.mode`.
Note that case b) is usually covered by case c).

BUG=b:177877310
TEST=Verified wireguard command on the device in the above cases.

Change-Id: I90cbb670a8149a27dc24ee97d9d8e1fe1c454d6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2967858
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>