camera: Fix hotplug crashes in IP HAL
In the IP HAL the CameraDevice object is accessed by the HAL, but also
over Mojo by the IP peripheral service. The lifetime of the object was
being determined by the IP peripheral service. So it was possible to get
a use after free crash when a camera was hotplugged if the HAL tried to
access it after the object had been freed. The solution is to switch to
using a shared_ptr from a unique_ptr and track devices that are opended
by the HAL alongside the existing lifetime management.

BUG=b:189462897
TEST=Manually open/closing camera, as well as hotplugging

Change-Id: I372a4227a94691d86c3898e753d498bdf2aa287b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2971564
Commit-Queue: Pablo Ceballos <pceballos@chromium.org>
Tested-by: Pablo Ceballos <pceballos@chromium.org>
Reviewed-by: Shik Chen <shik@chromium.org>