os_install_service: ensure log lines are written fully
When writing to the log file, use write_all instead of write. Calling
write_all ensures that all the input data is written, whereas write
acts like the C function of the same name; it returns the amount
written and it's up to the caller to call it again if necessary.

In practice this isn't likely to matter since the amount of data being
written is small.

BUG=None
TEST=cros_workon_make --test os_install_service

Change-Id: I86685238a815312d72b418f239ebba505ab37c09
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2975020
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>