os_install_service: stream chromeos-install output to the logger
Previously output from the chromeos-install command was captured in
memory and sent to the logger all at once after it either finished or
failed. This command can take a while to run though, so it's useful to
see output immediately. Added a new run_command_log_output function
that runs the command with stderr redirected to stdout, and copies
lines from stdout to the logger until the command exits.

BUG=None
TEST=cros_workon_make --install --test os_install_service
     ./build_image --board=reven test
     Run the image and "tail -f /var/log/messages"
     Start the install through the UI
     Observe that lines from chromeos-install are immediately logged
     rather than all at once at the end.

Change-Id: Idd7b389045fc79fc9ee0fb0324655ea13224d088
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2976180
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>