sepolicy: allow chromeos-install to run udevadm settle
The chromeos-install script calls "udevadm settle" before re-reading
the partition table. Since the partition table has just been modified
this settle operation is needed in case udev is in the middle of
re-reading the partition table, otherwise the call to blockdev can
fail. In order for the settle command to work it needs the watch
permission so it can monitor changes in /run/udev.

BUG=b:182386612
TEST=cros_workon_make --install --test selinux-policy
     ./build_image --board=reven base
     Run the image and start the installer through the UI, wait
     for it to complete successfully. Repeat this 5x since this
     fix is for a race condition.

Change-Id: I71548d5e5bdce8c426a2b162e00ac91b99e5e83c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2976285
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>