shill: vpn: Set kCredential for keys in WireGuardDriver
A property marked with kCredential will be cleared when
Service::Unload() is called. This patch applies this rule to PrivateKey
and PresharedKey in WireGuardDriver, to keep the behavior same between
all VPNDrivers.

After this patch, "SaveCredentials" needs to be set to true for a wg
service, since it doesn't make sense to let users input the keys every
time they want to connect to the service.

Also note that, if there is already a wg service in the profile, before
applying this patch, the "WireGuard.PrivateKey" field needs to be
removed (or add a prefix with "Credential." to the key), otherwise shill
will crash and restart without user profile. The reason is that
VPNDriver will try to read the field without the prefix in ROT47, but a
plain text is given so shill will crash. See VPNDriver::Load() and
VPNDriver::MigrateDeprecatedStorage() for more details.

BUG=b:177876632
TEST=unit_tests

Change-Id: Ib2eae6527109250e86af628f461ef725646b9475
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2978166
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>