cryptohome: Fix Keyset management unit test for resaving LE Credentials
These are the following three things that are changed about this test
in the CL:

1. The test ReSaveOnLoadTestLeCreds currently only checks a credential
should be resaved and does not follow through with calling resave on it.
This CL fixes that.

2. The test is setup with regular password credentials and not
LECredentials. This CL addresses that by setting  up the test with right
credentials. As a result, there is a need for using FakeLeCredentials as
opposed to current MockLECredentials to keep the test more readable.

3. reset_seed for LECredential is not stored as part of the keyset, but
the seed is used to generate key material in the keyset. If the
reset_seed is missing, resave should not go through. This is the expected
behavior and should have been part of this test.

BUG=chromium:1220871
TEST=FEATURES=test emerge-${zork} cryptohome

Change-Id: Ia0536bd7dc93b462e9a017c502b9b7e5cc76528d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2990192
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>
Auto-Submit: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>