rmad: Notify if device needs to keep open after disabling HWWP
Ask the repair tech if they want to enable factory mode if HWWP is off
and factory mode is currently disabled. This only happens when the
device is not enrolled, and the repair tech chooses to disable HWWP
physically. If they choose to enable factory mode, the system will
reboot and return to the same state.

Also inform the repair tech if they need to keep the device open after
disabling HWWP. This happens when cr50 factory mode is disabled. If the
device is not enrolled, the repair tech can choose to enable factory
mode (described in the section above), and after the reboot, the system
will notify that they can close the device. If the device is enrolled,
they cannot enable cr50 factory mode, and the device must be left open
throughout the repair.

BUG=b:189124401
TEST=emerge-volteer chromeos-base/system_api rmad;
     cros_run_unit_tests --board volteer --packages rmad

Change-Id: I364e8ebe1a2a905c6cbf53947c74a06a1d1c9230
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2994781
Tested-by: Cheng-Han Yang <chenghan@chromium.org>
Reviewed-by: Gene Chang <genechang@google.com>
Commit-Queue: Cheng-Han Yang <chenghan@chromium.org>