vm_tools: concierge: refuse to perform export when VM is running
If a VM is running it is quite likely that the exported disk image will
not be consistent; to prevent that we want to make sure that the VM is
stopped (not running) when we initiate export operation.

For Parallels VMs there is additional constraint: the VM needs to be
shut down, not merely stopped (which for Parallels means the VM is
suspended).

The caller may override this behavior by setting "force" flag in the
export request to "true".

BUG=b:184565415
TEST=Try exporting disk images for VMs that are running.

Change-Id: I84b9c3a2e7af37c00e3deced882038a6694b2753
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2995582
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Tested-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Queue: Dmitry Torokhov <dtor@chromium.org>