arc: Fix Platform.Memory.ARC.* UMA recording
To record the metrics correctly, metrics_daemon needs to read
/run/containers/android-run_oci/container.pid file to know init
process of the ARC++ container. Currently, this is not possible
because 1) the file is not bind mounted into the daemon's mount
namespace at all, and 2) SELinux policy to allow the daemon to
read the file is not available. This CL fixes both problems.

BUG=b:192035363
TEST=On octopus (which is with ARC++ P container), sign in and
 wait for 10 minutes. Visit chrome://histograms and verify
 Platform.Memory.ARC.Total is not 0.

Change-Id: I4e54cd60bd38c17d7527d23cdc214d8e287ba49e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2987640
Tested-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Yusuke Sato <yusukes@chromium.org>