vm_tools: concierge: always initialise topology properly
We need to always setup topology and initialize NumCPUs()
so that when we create VmBuilder we always pass correct
number of CPUs to it.

This fixes a regression on boards for which we have manually
disabled request.enable_rt_vcpu(). On such board we would
use topology with 0 CPUs and 0 RT CPUs.

BUG=b:192019034
TEST=arc.Boot.vm passes on 2 CPU akemi
TEST=arc.Boot.vm passes on kukui-arc-r
TEST=unit test

Disallow-Recycled-Builds: test-failures
Change-Id: I244d90a72ebbc149d282c49204ab746f257118ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2988004
Commit-Queue: Sergey Senozhatsky <senozhatsky@chromium.org>
Tested-by: Sergey Senozhatsky <senozhatsky@chromium.org>
Auto-Submit: Sergey Senozhatsky <senozhatsky@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>