cryptohome: trunks: don't wait for the handle flushed
We didn't need to wait for the handle flushed, because afte
we sended the flush request to trunks, it would be queued
in the dbus of trunks, and we didn't care about the result.
This could also reduce the latency of mount operation in
cryptohome.

BUG=b:188742402
TEST=unittest of trunks & cryptohome
TEST=tast run $DUT hwsec.*

Change-Id: I810bb464def5091a7ed53b993c132f87039e7ebd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2988209
Tested-by: Yi Chou <yich@google.com>
Commit-Queue: Yi Chou <yich@google.com>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>