dev-install: suppress bzip2 warnings due to xpaks
The install process shows a lot of harmless, but confusing, warnings
due to the format of Gentoo binpkgs -- an xpak file is appended to
the bz2 file, and bzip2 warns about the unused data.  Pass the -q
flag to bzip2 to suppress that and tighten up the logs.

BUG=None
TEST=CQ passes

Change-Id: I92fbc087891daf041ed2f6e7d608fba8a71951ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2994735
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>