vm_tools: crostini_client: add support for "force" flag for export operation
Concierge now refused to perform export operation is VM is running or is
not fully shut down in case of parallels. However if the system is in a
bad state we still need to provide users with the ability to extract
their data from their device. In this case they can use -f or --force
option to bypass checks in concierge:

	vmc export -f PvmDefault VM.zip

BUG=b:184565415
TEST=Try exporting disk images for VMs that are running, then do the
same with "-f" flag..

Change-Id: Ib0c43b7123f3559aecbcba19db724cc5f6eb7a6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2995583
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Tested-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Queue: Dmitry Torokhov <dtor@chromium.org>