libhwsec-foundation: add helpers for runtime TPM selection
Add helpers to switching between build-time code path and run-time code
path on the caller side.

When tpm_dynamic is not enabled, the codes inside *_SECTION shuold be
decided in the build-time, and unrelated *_SECTION should not affect the
build result.

When tpm_dynamic is not enabled, the codes inside *_SECTION shuold be
decided in the run-time, and all *_SECTION would be compiled into the
same executable.

BUG=b:179811538
TEST=Build OK.

Cq-Depend: chromium:2995121
Change-Id: I6c0962b8f068b79bc6093e0cd87fd657af4d018d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2993960
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>
Tested-by: Yi Chou <yich@google.com>
Commit-Queue: Yi Chou <yich@google.com>