camera: enable HDRnet for all resolutions
This CL allows HDRnet stream manipulator to work even if the stream
configuration asks for YUV streams with different aspect ratios. It
should be rare to request YUV streams with different aspect ratios in
practice, so we can accept some performance overhead in this case.

While we're at it, also renames a few utility methods for better
sementics, and remove some redundant uniforms in the shaders.

BUG=b:187157014
TEST=hdrnet_stream_manipulator_test

Change-Id: I5fc986fd7e4ef914408b9c70bb85ec920d6860d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2997428
Tested-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Wei Lee <wtlee@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>