cryptohome: Add HSM payload generation
This change adds generation of HSM payload that will happen at
onboarding. The data consists of associated data and encrypted plain
text and contains all the info needed to perform migration. The CL
does not take into account the 'epoch' part of the protocol that will
be implemented in the next step.
The old 'toy' protocol was left for reference but will have to be
removed in the final implementation.

BUG=b:192786922
TEST=cros_run_unit_tests --board=${BOARD} --packages cryptohome

Cq-Depend: chromium:3065663
Change-Id: Iaf0f452ebe782f262a707b30c6dc8446d0d600e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3001857
Tested-by: Maciek Slusarczyk <mslus@chromium.org>
Commit-Queue: Maciek Slusarczyk <mslus@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>