missive: Ensure upload retry on failure
This fixes a subtle problem: for IMMEDIATE and MANUAL queues if upload
fails for whatever reason, we are not redoing it until the next event is
enqueued or next Flush is called (for other queues we just do next
periodic upload, but these two have none).
Also fixes a minor inefficiency in non-encrypted storage (which is now
only used in tests).

BUG=b:192666219
TEST=FEATURES="test" emerge-$BOARD missive

Change-Id: Id7140e22a82a8881f285280cd0d1e2f5b541aa9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3003051
Reviewed-by: Kyle Williams <kdgwill@chromium.org>
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>