patchpanel: Fix DNS rule for IPv6
IPv6 rule for DNS forwarding calls iptables with
"addr:port" as its argument. This is incorrect as the
port will be treated as part of the IPv6 address.
This patch changed the argument to be "addr" only
as there is no need to change the port.

BUG=b:180551553
TEST=ip6tables -S;
     # tested alongside DNS proxy IPv6 patch.
TEST=units

Change-Id: Id533e316ed0d59bd0aba9c50ab4cc6f878427795
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3009919
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Garrick Evans <garrick@chromium.org>