dlcservice: Fix DLC installation deadlock
When the OS is updated, the update_engine daemon goes into a need reboot
state. The dlcservice daemon realizes this and holds back all incoming
installation requests.

This means that any new request to install a DLC when the device is
updated does not cancel the installation.
Leaving the DLC in a deadlock state of "installing" when in reality, no
daemon is doing the installation at all.

BUG=chromium:1225915
TEST=FEATURES=test emerge-$B dlcservice

Change-Id: Iaa656bec08636a95e1665c9861b8683bd9f9be78
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3011939
Tested-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Vyshu Khota <vyshu@chromium.org>