missive: Further ensure upload retry on failure
Copy two changes from Chromium:
If upload succeeded but during retry delay period not all events have been
confirmed, retry the upload, as if it failed.
Also, Fix WeakPtr usage in ReadContext: Factory was mistakenly placed in
ReadContext, which made it useless. It deactivated weak pointers not when
StorageQueue was destructed but when ReadContext was.

BUG=b:192666219
TEST=FEATURES="test" emerge-$BOARD missive

Change-Id: Ie5412845be0ae52824163b21d4c2c8ab8de51701
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3012353
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Reviewed-by: Kyle Williams <kdgwill@chromium.org>
Reviewed-by: Zach Trudo <zatrudo@google.com>