login: Rename kBrowserDataMigrationFlag.
This patch renames the flag to `kBrowserDataMigrationForUserFlag` to
make it more in line with the implementation. Since this flag is not in
use yet, there will be no impact to the product.

http://go/lacros-data-migration-v2.

BUG=chromium:1178702
TEST=None

Change-Id: Ie1f9df440e5096520a360c3135116d764183da63
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3012364
Commit-Queue: Yuta Hijikata <ythjkt@chromium.org>
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Yuta Hijikata <ythjkt@chromium.org>
Auto-Submit: Yuta Hijikata <ythjkt@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>